backend/requirements.txt
flask
flask-cors
weasyprint
google-genai
python-dotenv
